주차장 위도 경도 소수점으로 값 받아옴. 쿼리스트링 \n오류 수정

+++ carpark/src/com/carpark/admin/model/ParkingDto.java
@@ -6,8 +6,8 @@ public class ParkingDto {
 	private String location;
 	private int park_capacity;
 	private String owner_id;
-	private int latitude;
-	private int longitude;
+	private double latitude;
+	private double longitude;
 	private String park_type;
 	
 	
@@ -41,16 +41,16 @@ public class ParkingDto {
 	public void setOwner_id(String owner_id) {
 		this.owner_id = owner_id;
 	}
-	public int getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(int latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
-	public int getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(int longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 	public String getPark_type() {
